port to windows

+++ Demo/Program/tri_demo.c
@@ -229,7 +229,7 @@ int main (int argc, char **argv)
 
         double t_mark [2] = { 0, 0 } ;
         OK (tricount (&ntri1 [nthreads], 3, NULL, NULL, L, NULL, t_mark)) ;
-        printf ("triangles, method 3: %ld\n", ntri1 [nthreads]) ;
+        printf ("triangles, method 3: %g\n", (double) ntri1 [nthreads]) ;
         if (ntri1 [nthreads] != nt)
         {
             printf ("error 3!\n") ;
